refactor: move feed assets to dedicated module
