+++ frontend/__init__.py
@@ -1,5 +1,14 @@
 """Convenience exports for frontend utilities."""
 
 from .theme import apply_theme, inject_modern_styles, get_accent_color
+from .assets import story_css, story_js, reaction_css, scroll_js
 
-__all__ = ["apply_theme", "inject_modern_styles", "get_accent_color"]
+__all__ = [
+    "apply_theme",
+    "inject_modern_styles",
+    "get_accent_color",
+    "story_css",
+    "story_js",
+    "reaction_css",
+    "scroll_js",
+]
