+++ transcendental_resonance_frontend/pages/feed.py
@@ -15,6 +15,7 @@ from frontend.theme import apply_theme
 from modern_ui import inject_modern_styles
 from streamlit_helpers import theme_selector, safe_container, sanitize_text
 from modern_ui_components import st_javascript
+from frontend.assets import story_css, story_js, reaction_css, scroll_js
 
 # ──────────────────────────────────────────────────────────────────────────────
 # Sample data models
@@ -79,74 +80,10 @@ def _generate_posts(count: int, start: int = 0) -> List[Post]:
 # Rendering helpers
 # ──────────────────────────────────────────────────────────────────────────────
 
-_STORY_CSS = """
-<style>
-.story-strip{display:flex;overflow-x:auto;gap:0.5rem;padding:0.5rem;margin-bottom:1rem;}
-.story-item{flex:0 0 auto;text-align:center;font-size:0.8rem;color:var(--text-muted);}
-.story-item img{border-radius:50%;border:2px solid var(--accent);}
-.post-card{background:var(--card);padding:0.5rem 0;border-radius:12px;
-           margin-bottom:1rem;box-shadow:0 1px 2px rgba(0,0,0,0.05);}
-.post-header{display:flex;align-items:center;gap:0.5rem;padding:0 0.5rem;margin-bottom:0.5rem;}
-.post-header img{border-radius:50%;width:40px;height:40px;}
-.post-caption{padding:0.25rem 0.5rem;}
-</style>
-"""
-
-_STORY_JS = """
-(() => {
-  const strip = document.getElementById('story-strip');
-  if (!strip || window.storyCarouselInit) return;
-  window.storyCarouselInit = true;
-  let idx = 0;
-  const advance = () => {
-    idx = (idx + 1) % strip.children.length;
-    const el = strip.children[idx];
-    strip.scrollTo({left: el.offsetLeft, behavior: 'smooth'});
-  };
-  let interval = setInterval(advance, 3000);
-  let startX = 0;
-  let scrollLeft = 0;
-  strip.addEventListener('touchstart', (e) => {
-    clearInterval(interval);
-    startX = e.touches[0].pageX;
-    scrollLeft = strip.scrollLeft;
-  });
-  strip.addEventListener('touchmove', (e) => {
-    const x = e.touches[0].pageX;
-    const walk = startX - x;
-    strip.scrollLeft = scrollLeft + walk;
-  });
-  strip.addEventListener('touchend', () => {
-    interval = setInterval(advance, 3000);
-  });
-})();
-"""
-
-_REACTION_CSS = """
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css">
-<style>
-.reaction-btn{background:transparent;border:none;font-size:1.1rem;cursor:pointer;margin-right:0.25rem;transition:transform 0.1s ease;}
-.reaction-btn:active{transform:scale(1.2);}
-</style>
-"""
-
-_SCROLL_JS = """
-<script>
-const sentinel = document.getElementById('load-sentinel');
-if(sentinel){
-  const observer = new IntersectionObserver((entries)=>{
-    entries.forEach(e=>{if(e.isIntersecting){const btn=document.getElementById('load-more-btn');btn&&btn.click();}});
-  });
-  observer.observe(sentinel);
-}
-</script>
-"""
-
-
 
 def _render_stories(users: List[User]) -> None:
     """Render the horizontal story-strip."""
-    st.markdown(_STORY_CSS, unsafe_allow_html=True)
+    st.markdown(story_css(), unsafe_allow_html=True)
     html = "<div class='story-strip' id='story-strip'>"
     for u in users:
         avatar = sanitize_text(u.avatar)
@@ -156,7 +93,7 @@ def _render_stories(users: List[User]) -> None:
         )
     html += "</div>"
     st.markdown(html, unsafe_allow_html=True)
-    st_javascript(_STORY_JS, key="story_carousel")
+    st_javascript(story_js(), key="story_carousel")
 
 
 def _render_post(post: Post) -> None:
@@ -285,7 +222,7 @@ def _page_body() -> None:
     users = _sample_users()
 
     _render_stories(users)
-    st.markdown(_REACTION_CSS, unsafe_allow_html=True)
+    st.markdown(reaction_css(), unsafe_allow_html=True)
 
     offset = st.session_state["post_offset"]
     for p in posts[:offset]:
